chore: select refactor
