+++ src/ui/canvas/tools/improved-select/types.ts
@@ -1,0 +1,12 @@
+export type ImageHandleMode =
+  | 'rotate'
+  | 'tl'
+  | 'tr'
+  | 'bl'
+  | 'br'
+  | 'crop-t'
+  | 'crop-b'
+  | 'crop-l'
+  | 'crop-r'
+  | 'move'
+  | null;
